feat: Update useConnectDiscord
