+++ apps/gamification/views/Profile/hooks/settingsModal/useConnectDiscord.tsx
@@ -20,29 +20,38 @@ interface DiscordResponse {
 
 export const useConnectDiscord = () => {
   useEffect(() => {
-    // Url will be like #token_type=Bearer&access_token=${access_token}&expires_in=${expires_in}&scope=identify
-    const fragment = new URLSearchParams(window.location.hash.slice(1))
-    const accessToken = fragment.get('access_token')
+    const fetchUser = async () => {
+      // Url will be like #token_type=Bearer&access_token=${access_token}&expires_in=${expires_in}&scope=identify
+      const fragment = new URLSearchParams(window.location.hash.slice(1))
+      const accessToken = fragment.get('access_token')
 
-    if (accessToken) {
-      fetch('https://discord.com/api/users/@me', {
-        headers: {
-          authorization: `Bearer ${accessToken}`,
-        },
-      })
-        .then((result) => result.json())
-        .then((response: DiscordResponse) => {
+      // Update the URL in the address bar without reloading the page
+      const newURL = `${window.location.origin}${window.location.pathname}`
+      window.history.pushState({}, '', newURL)
+
+      if (accessToken) {
+        try {
+          const response = await fetch('https://discord.com/api/users/@me', {
+            headers: {
+              authorization: `Bearer ${accessToken}`,
+            },
+          })
+          const data: DiscordResponse = await response.json()
           // response.id is what we want
-          console.log('response', response)
-        })
-        .catch(console.error)
+          console.log(data)
+        } catch (error) {
+          console.error('Error fetching discord user data:', error)
+        }
+      }
     }
+
+    fetchUser()
   }, [])
 
   const connect = () => {
     const url =
       'https://discord.com/oauth2/authorize?client_id=1247444580874453065&response_type=token&redirect_uri=http%3A%2F%2Flocalhost%3A3005%2Fprofile&scope=identify'
-    window.open(url)
+    window.location.href = url
   }
 
   return {
